refactor(backend): Refactor and fix multiple issues
