+++ Backend/Program.cs
@@ -158,8 +158,6 @@ builder.Services.AddScoped<ResourceService>();
 builder.Services.AddScoped<CourseModuleService>();
 builder.Services.AddScoped<CourseTagService>();
 builder.Services.AddScoped<ModuleResourceService>();
-builder.Services.AddScoped<CourseRegistrationService>();
-
 
 // Build App
 var app = builder.Build();
