+++ Backend/Services/ModuleService.cs
@@ -1,7 +1,8 @@
 using Backend.Context;
 using Backend.Dtos;
-using Microsoft.EntityFrameworkCore;
+using Backend.Dtos.Module;
 using Backend.Models;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
 
 namespace Backend.Services
